Ejected minify logic from Widget view
